Network.AWS.CodeDeploy

Description

Amazon CodeDeploy is a service that automates code deployments to Amazon EC2 instances. Amazon CodeDeploy makes it easier for you to rapidly release new features, helps you avoid downtime during deployment, and handles the complexity of updating your applications. You can use Amazon CodeDeploy to automate deployments, eliminating the need for error-prone manual operations, and the service scales with your infrastructure so you can easily deploy to one EC2 instance or thousands.

Documentation

module Network.AWS.CodeDeploy.AddTagsToOnPremisesInstances

module Network.AWS.CodeDeploy.BatchGetApplications

module Network.AWS.CodeDeploy.BatchGetDeployments

module Network.AWS.CodeDeploy.BatchGetOnPremisesInstances

module Network.AWS.CodeDeploy.CreateApplication

module Network.AWS.CodeDeploy.CreateDeployment

module Network.AWS.CodeDeploy.CreateDeploymentConfig

module Network.AWS.CodeDeploy.CreateDeploymentGroup

module Network.AWS.CodeDeploy.DeleteApplication

module Network.AWS.CodeDeploy.DeleteDeploymentConfig

module Network.AWS.CodeDeploy.DeleteDeploymentGroup

module Network.AWS.CodeDeploy.DeregisterOnPremisesInstance

module Network.AWS.CodeDeploy.GetApplication

module Network.AWS.CodeDeploy.GetApplicationRevision

module Network.AWS.CodeDeploy.GetDeployment

module Network.AWS.CodeDeploy.GetDeploymentConfig

module Network.AWS.CodeDeploy.GetDeploymentGroup

module Network.AWS.CodeDeploy.GetDeploymentInstance

module Network.AWS.CodeDeploy.GetOnPremisesInstance

module Network.AWS.CodeDeploy.ListApplicationRevisions

module Network.AWS.CodeDeploy.ListApplications

module Network.AWS.CodeDeploy.ListDeploymentConfigs

module Network.AWS.CodeDeploy.ListDeploymentGroups

module Network.AWS.CodeDeploy.ListDeploymentInstances

module Network.AWS.CodeDeploy.ListDeployments

module Network.AWS.CodeDeploy.ListOnPremisesInstances

module Network.AWS.CodeDeploy.RegisterApplicationRevision

module Network.AWS.CodeDeploy.RegisterOnPremisesInstance

module Network.AWS.CodeDeploy.RemoveTagsFromOnPremisesInstances

module Network.AWS.CodeDeploy.StopDeployment

module Network.AWS.CodeDeploy.Types

module Network.AWS.CodeDeploy.UpdateApplication

module Network.AWS.CodeDeploy.UpdateDeploymentGroup