Network.AWS.CodeDeploy.AddTagsToOnPremisesInstances

Description

Adds tags to on-premises instances.

Creating a Request

addTagsToOnPremisesInstances :: AddTagsToOnPremisesInstances

Creates a value of AddTagsToOnPremisesInstances with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

attopiTags - The tag key-value pairs to add to the on-premises instances. Keys and values are both required. Keys cannot be null or empty strings. Value-only tags are not allowed.

attopiInstanceNames - The names of the on-premises instances to which to add tags.

data AddTagsToOnPremisesInstances

Represents the input of, and adds tags to, an on-premises instance operation.

See: addTagsToOnPremisesInstances smart constructor.

Request Lenses

attopiTags :: Lens' AddTagsToOnPremisesInstances [Tag]

The tag key-value pairs to add to the on-premises instances. Keys and values are both required. Keys cannot be null or empty strings. Value-only tags are not allowed.

attopiInstanceNames :: Lens' AddTagsToOnPremisesInstances [Text]

The names of the on-premises instances to which to add tags.

Destructuring the Response

addTagsToOnPremisesInstancesResponse :: AddTagsToOnPremisesInstancesResponse

Creates a value of AddTagsToOnPremisesInstancesResponse with the minimum fields required to make a request.

data AddTagsToOnPremisesInstancesResponse

See: addTagsToOnPremisesInstancesResponse smart constructor.