Network.AWS.CodeDeploy.BatchGetApplicationRevisions

Description

Gets information about one or more application revisions.

Creating a Request

batchGetApplicationRevisions

Arguments

:: Text
bgarApplicationName

-> BatchGetApplicationRevisions

Creates a value of BatchGetApplicationRevisions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bgarApplicationName - The name of an AWS CodeDeploy application about which to get revision information.

bgarRevisions - Information to get about the application revisions, including type and location.

data BatchGetApplicationRevisions

Represents the input of a BatchGetApplicationRevisions operation.

See: batchGetApplicationRevisions smart constructor.

Request Lenses

bgarApplicationName :: Lens' BatchGetApplicationRevisions Text

The name of an AWS CodeDeploy application about which to get revision information.

bgarRevisions :: Lens' BatchGetApplicationRevisions [RevisionLocation]

Information to get about the application revisions, including type and location.

Destructuring the Response

batchGetApplicationRevisionsResponse

Arguments

:: Int
bgarrsResponseStatus

-> BatchGetApplicationRevisionsResponse

Creates a value of BatchGetApplicationRevisionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bgarrsApplicationName - The name of the application that corresponds to the revisions.

bgarrsRevisions - Additional information about the revisions, including the type and location.

bgarrsErrorMessage - Information about errors that may have occurred during the API call.

bgarrsResponseStatus - -- | The response status code.

data BatchGetApplicationRevisionsResponse

Represents the output of a BatchGetApplicationRevisions operation.

See: batchGetApplicationRevisionsResponse smart constructor.

Response Lenses

bgarrsApplicationName :: Lens' BatchGetApplicationRevisionsResponse (Maybe Text)

The name of the application that corresponds to the revisions.

bgarrsRevisions :: Lens' BatchGetApplicationRevisionsResponse [RevisionInfo]

Additional information about the revisions, including the type and location.

bgarrsErrorMessage :: Lens' BatchGetApplicationRevisionsResponse (Maybe Text)

Information about errors that may have occurred during the API call.

bgarrsResponseStatus :: Lens' BatchGetApplicationRevisionsResponse Int

- | The response status code.