Network.AWS.CodeDeploy.BatchGetDeploymentInstances

Description

Gets information about one or more instance that are part of a deployment group.

Creating a Request

batchGetDeploymentInstances

Arguments

:: Text
bgdiDeploymentId

-> BatchGetDeploymentInstances

Creates a value of BatchGetDeploymentInstances with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bgdiDeploymentId - The unique ID of a deployment.

bgdiInstanceIds - The unique IDs of instances in the deployment group.

data BatchGetDeploymentInstances

Represents the input of a BatchGetDeploymentInstances operation.

See: batchGetDeploymentInstances smart constructor.

Request Lenses

bgdiDeploymentId :: Lens' BatchGetDeploymentInstances Text

The unique ID of a deployment.

bgdiInstanceIds :: Lens' BatchGetDeploymentInstances [Text]

The unique IDs of instances in the deployment group.

Destructuring the Response

batchGetDeploymentInstancesResponse

Arguments

:: Int
bgdirsResponseStatus

-> BatchGetDeploymentInstancesResponse

Creates a value of BatchGetDeploymentInstancesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bgdirsInstancesSummary - Information about the instance.

bgdirsErrorMessage - Information about errors that may have occurred during the API call.

bgdirsResponseStatus - -- | The response status code.

data BatchGetDeploymentInstancesResponse

Represents the output of a BatchGetDeploymentInstances operation.

See: batchGetDeploymentInstancesResponse smart constructor.

Response Lenses

bgdirsInstancesSummary :: Lens' BatchGetDeploymentInstancesResponse [InstanceSummary]

Information about the instance.

bgdirsErrorMessage :: Lens' BatchGetDeploymentInstancesResponse (Maybe Text)

Information about errors that may have occurred during the API call.

bgdirsResponseStatus :: Lens' BatchGetDeploymentInstancesResponse Int

- | The response status code.