Network.AWS.CodeDeploy.ContinueDeployment

Description

For a blue/green deployment, starts the process of rerouting traffic from instances in the original environment to instances in the replacement environment without waiting for a specified wait time to elapse. (Traffic rerouting, which is achieved by registering instances in the replacement environment with the load balancer, can start as soon as all instances have a status of Ready.)

Creating a Request

continueDeployment :: ContinueDeployment

Creates a value of ContinueDeployment with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdDeploymentId - The deployment ID of the blue/green deployment for which you want to start rerouting traffic to the replacement environment.

data ContinueDeployment

See: continueDeployment smart constructor.

Request Lenses

cdDeploymentId :: Lens' ContinueDeployment (Maybe Text)

The deployment ID of the blue/green deployment for which you want to start rerouting traffic to the replacement environment.

Destructuring the Response

continueDeploymentResponse :: ContinueDeploymentResponse

Creates a value of ContinueDeploymentResponse with the minimum fields required to make a request.

data ContinueDeploymentResponse

See: continueDeploymentResponse smart constructor.