Network.AWS.CodeDeploy.CreateApplication

Description

Creates an application.

Creating a Request

createApplication

Arguments

:: Text
caApplicationName

-> CreateApplication

Creates a value of CreateApplication with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

caApplicationName - The name of the application. This name must be unique with the applicable IAM user or AWS account.

data CreateApplication

Represents the input of a CreateApplication operation.

See: createApplication smart constructor.

Request Lenses

caApplicationName :: Lens' CreateApplication Text

The name of the application. This name must be unique with the applicable IAM user or AWS account.

Destructuring the Response

createApplicationResponse

Arguments

:: Int
carsResponseStatus

-> CreateApplicationResponse

Creates a value of CreateApplicationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

carsApplicationId - A unique application ID.

carsResponseStatus - -- | The response status code.

data CreateApplicationResponse

Represents the output of a CreateApplication operation.

See: createApplicationResponse smart constructor.

Response Lenses

carsApplicationId :: Lens' CreateApplicationResponse (Maybe Text)

A unique application ID.

carsResponseStatus :: Lens' CreateApplicationResponse Int

- | The response status code.