Network.AWS.CodeDeploy.CreateDeployment

Description

Deploys an application revision through the specified deployment group.

Creating a Request

createDeployment

Arguments

:: Text
cdApplicationName

-> CreateDeployment

Creates a value of CreateDeployment with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdDeploymentConfigName - The name of a deployment configuration associated with the applicable IAM user or AWS account. If not specified, the value configured in the deployment group will be used as the default. If the deployment group does not have a deployment configuration associated with it, then CodeDeployDefault.OneAtATime will be used by default.

cdFileExistsBehavior - Information about how AWS CodeDeploy handles files that already exist in a deployment target location but weren't part of the previous successful deployment. The fileExistsBehavior parameter takes any of the following values: * DISALLOW: The deployment fails. This is also the default behavior if no option is specified. * OVERWRITE: The version of the file from the application revision currently being deployed replaces the version already on the instance. * RETAIN: The version of the file already on the instance is kept and used as part of the new deployment.

cdTargetInstances - Information about the instances that will belong to the replacement environment in a blue/green deployment.

cdRevision - The type and location of the revision to deploy.

cdDescription - A comment about the deployment.

cdAutoRollbackConfiguration - Configuration information for an automatic rollback that is added when a deployment is created.

cdUpdateOutdatedInstancesOnly - Indicates whether to deploy to all instances or only to instances that are not running the latest application revision.

cdDeploymentGroupName - The name of the deployment group.

cdIgnoreApplicationStopFailures - If set to true, then if the deployment causes the ApplicationStop deployment lifecycle event to an instance to fail, the deployment to that instance will not be considered to have failed at that point and will continue on to the BeforeInstall deployment lifecycle event. If set to false or not specified, then if the deployment causes the ApplicationStop deployment lifecycle event to fail to an instance, the deployment to that instance will stop, and the deployment to that instance will be considered to have failed.

cdApplicationName - The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.

data CreateDeployment

Represents the input of a CreateDeployment operation.

See: createDeployment smart constructor.

Request Lenses

cdDeploymentConfigName :: Lens' CreateDeployment (Maybe Text)

The name of a deployment configuration associated with the applicable IAM user or AWS account. If not specified, the value configured in the deployment group will be used as the default. If the deployment group does not have a deployment configuration associated with it, then CodeDeployDefault.OneAtATime will be used by default.

cdFileExistsBehavior :: Lens' CreateDeployment (Maybe FileExistsBehavior)

Information about how AWS CodeDeploy handles files that already exist in a deployment target location but weren't part of the previous successful deployment. The fileExistsBehavior parameter takes any of the following values: * DISALLOW: The deployment fails. This is also the default behavior if no option is specified. * OVERWRITE: The version of the file from the application revision currently being deployed replaces the version already on the instance. * RETAIN: The version of the file already on the instance is kept and used as part of the new deployment.

cdTargetInstances :: Lens' CreateDeployment (Maybe TargetInstances)

Information about the instances that will belong to the replacement environment in a blue/green deployment.

cdRevision :: Lens' CreateDeployment (Maybe RevisionLocation)

The type and location of the revision to deploy.

cdDescription :: Lens' CreateDeployment (Maybe Text)

A comment about the deployment.

cdAutoRollbackConfiguration :: Lens' CreateDeployment (Maybe AutoRollbackConfiguration)

Configuration information for an automatic rollback that is added when a deployment is created.

cdUpdateOutdatedInstancesOnly :: Lens' CreateDeployment (Maybe Bool)

Indicates whether to deploy to all instances or only to instances that are not running the latest application revision.

cdDeploymentGroupName :: Lens' CreateDeployment (Maybe Text)

The name of the deployment group.

cdIgnoreApplicationStopFailures :: Lens' CreateDeployment (Maybe Bool)

If set to true, then if the deployment causes the ApplicationStop deployment lifecycle event to an instance to fail, the deployment to that instance will not be considered to have failed at that point and will continue on to the BeforeInstall deployment lifecycle event. If set to false or not specified, then if the deployment causes the ApplicationStop deployment lifecycle event to fail to an instance, the deployment to that instance will stop, and the deployment to that instance will be considered to have failed.

cdApplicationName :: Lens' CreateDeployment Text

The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.

Destructuring the Response

createDeploymentResponse

Arguments

:: Int
cdrsResponseStatus

-> CreateDeploymentResponse

Creates a value of CreateDeploymentResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdrsDeploymentId - A unique deployment ID.

cdrsResponseStatus - -- | The response status code.

data CreateDeploymentResponse

Represents the output of a CreateDeployment operation.

See: createDeploymentResponse smart constructor.

Response Lenses

cdrsDeploymentId :: Lens' CreateDeploymentResponse (Maybe Text)

A unique deployment ID.

cdrsResponseStatus :: Lens' CreateDeploymentResponse Int

- | The response status code.