Network.AWS.CodeDeploy.CreateDeploymentConfig

Description

Creates a new deployment configuration.

http://docs.aws.amazon.com/codedeploy/latest/APIReference/API_CreateDeploymentConfig.html

Request

data CreateDeploymentConfig

Request constructor

createDeploymentConfig

Arguments

:: Text
cdcDeploymentConfigName

-> CreateDeploymentConfig

CreateDeploymentConfig constructor.

The fields accessible through corresponding lenses are:

cdcDeploymentConfigName :: Text

cdcMinimumHealthyHosts :: Maybe MinimumHealthyHosts

Request lenses

cdcDeploymentConfigName :: Lens' CreateDeploymentConfig Text

The name of the deployment configuration to create.

cdcMinimumHealthyHosts :: Lens' CreateDeploymentConfig (Maybe MinimumHealthyHosts)

The minimum number of healthy instances that should be available at any time during the deployment. There are two parameters expected in the input: type and value.

The type parameter takes either of the following values:

HOST_COUNT: The value parameter represents the minimum number of healthy instances, as an absolute value. FLEET_PERCENT: The value parameter represents the minimum number of healthy instances, as a percentage of the total number of instances in the deployment. If you specify FLEET_PERCENT, then at the start of the deployment AWS CodeDeploy converts the percentage to the equivalent number of instances and rounds fractional instances up. The value parameter takes an integer.

For example, to set a minimum of 95% healthy instances, specify a type of FLEET_PERCENT and a value of 95.

Response

data CreateDeploymentConfigResponse

Response constructor

createDeploymentConfigResponse :: CreateDeploymentConfigResponse

CreateDeploymentConfigResponse constructor.

The fields accessible through corresponding lenses are:

cdcrDeploymentConfigId :: Maybe Text

Response lenses

cdcrDeploymentConfigId :: Lens' CreateDeploymentConfigResponse (Maybe Text)

A unique deployment configuration ID.