Network.AWS.CodeDeploy.CreateDeploymentGroup

Description

Creates a new deployment group for application revisions to be deployed to.

http://docs.aws.amazon.com/codedeploy/latest/APIReference/API_CreateDeploymentGroup.html

Request

data CreateDeploymentGroup

Request constructor

createDeploymentGroup

Arguments

:: Text
cdgApplicationName

-> Text
cdgDeploymentGroupName

-> Text
cdgServiceRoleArn

-> CreateDeploymentGroup

CreateDeploymentGroup constructor.

The fields accessible through corresponding lenses are:

cdgApplicationName :: Text

cdgAutoScalingGroups :: [Text]

cdgDeploymentConfigName :: Maybe Text

cdgDeploymentGroupName :: Text

cdgEc2TagFilters :: [EC2TagFilter]

cdgOnPremisesInstanceTagFilters :: [TagFilter]

cdgServiceRoleArn :: Text

Request lenses

cdgApplicationName :: Lens' CreateDeploymentGroup Text

The name of an existing AWS CodeDeploy application associated with the applicable IAM user or AWS account.

cdgAutoScalingGroups :: Lens' CreateDeploymentGroup [Text]

A list of associated Auto Scaling groups.

cdgDeploymentConfigName :: Lens' CreateDeploymentGroup (Maybe Text)

If specified, the deployment configuration name must be one of the predefined values, or it can be a custom deployment configuration:

CodeDeployDefault.AllAtOnce deploys an application revision to up to all of the instances at once. The overall deployment succeeds if the application revision deploys to at least one of the instances. The overall deployment fails after the application revision fails to deploy to all of the instances. For example, for 9 instances, deploy to up to all 9 instances at once. The overall deployment succeeds if any of the 9 instances is successfully deployed to, and it fails if all 9 instances fail to be deployed to. CodeDeployDefault.HalfAtATime deploys to up to half of the instances at a time (with fractions rounded down). The overall deployment succeeds if the application revision deploys to at least half of the instances (with fractions rounded up); otherwise, the deployment fails. For example, for 9 instances, deploy to up to 4 instances at a time. The overall deployment succeeds if 5 or more instances are successfully deployed to; otherwise, the deployment fails. Note that the deployment may successfully deploy to some instances, even if the overall deployment fails. CodeDeployDefault.OneAtATime deploys the application revision to only one of the instances at a time. The overall deployment succeeds if the application revision deploys to all of the instances. The overall deployment fails after the application revision first fails to deploy to any one instances. For example, for 9 instances, deploy to one instance at a time. The overall deployment succeeds if all 9 instances are successfully deployed to, and it fails if any of one of the 9 instances fail to be deployed to. Note that the deployment may successfully deploy to some instances, even if the overall deployment fails. This is the default deployment configuration if a configuration isn't specified for either the deployment or the deployment group. To create a custom deployment configuration, call the create deployment configuration operation.

cdgDeploymentGroupName :: Lens' CreateDeploymentGroup Text

The name of an existing deployment group for the specified application.

cdgEc2TagFilters :: Lens' CreateDeploymentGroup [EC2TagFilter]

The Amazon EC2 tags to filter on.

cdgOnPremisesInstanceTagFilters :: Lens' CreateDeploymentGroup [TagFilter]

The on-premises instance tags to filter on.

cdgServiceRoleArn :: Lens' CreateDeploymentGroup Text

A service role ARN that allows AWS CodeDeploy to act on the user's behalf when interacting with AWS services.

Response

data CreateDeploymentGroupResponse

Response constructor

createDeploymentGroupResponse :: CreateDeploymentGroupResponse

CreateDeploymentGroupResponse constructor.

The fields accessible through corresponding lenses are:

cdgrDeploymentGroupId :: Maybe Text

Response lenses

cdgrDeploymentGroupId :: Lens' CreateDeploymentGroupResponse (Maybe Text)

A unique deployment group ID.