Network.AWS.CodeDeploy.DeleteDeploymentConfig

Description

Deletes a deployment configuration.

Creating a Request

deleteDeploymentConfig

Arguments

:: Text
ddcDeploymentConfigName

-> DeleteDeploymentConfig

Creates a value of DeleteDeploymentConfig with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddcDeploymentConfigName - The name of a deployment configuration associated with the applicable IAM user or AWS account.

data DeleteDeploymentConfig

Represents the input of a DeleteDeploymentConfig operation.

See: deleteDeploymentConfig smart constructor.

Request Lenses

ddcDeploymentConfigName :: Lens' DeleteDeploymentConfig Text

The name of a deployment configuration associated with the applicable IAM user or AWS account.

Destructuring the Response

deleteDeploymentConfigResponse :: DeleteDeploymentConfigResponse

Creates a value of DeleteDeploymentConfigResponse with the minimum fields required to make a request.

data DeleteDeploymentConfigResponse

See: deleteDeploymentConfigResponse smart constructor.