Network.AWS.CodeDeploy.DeleteDeploymentGroup

Description

Deletes a deployment group.

http://docs.aws.amazon.com/codedeploy/latest/APIReference/API_DeleteDeploymentGroup.html

Request

data DeleteDeploymentGroup

Request constructor

deleteDeploymentGroup

Arguments

:: Text
ddgApplicationName

-> Text
ddgDeploymentGroupName

-> DeleteDeploymentGroup

DeleteDeploymentGroup constructor.

The fields accessible through corresponding lenses are:

ddgApplicationName :: Text

ddgDeploymentGroupName :: Text

Request lenses

ddgApplicationName :: Lens' DeleteDeploymentGroup Text

The name of an existing AWS CodeDeploy application associated with the applicable IAM user or AWS account.

ddgDeploymentGroupName :: Lens' DeleteDeploymentGroup Text

The name of an existing deployment group for the specified application.

Response

data DeleteDeploymentGroupResponse

Response constructor

deleteDeploymentGroupResponse :: DeleteDeploymentGroupResponse

DeleteDeploymentGroupResponse constructor.

The fields accessible through corresponding lenses are:

ddgrHooksNotCleanedUp :: [AutoScalingGroup]

Response lenses

ddgrHooksNotCleanedUp :: Lens' DeleteDeploymentGroupResponse [AutoScalingGroup]

If the output contains no data, and the corresponding deployment group contained at least one Auto Scaling group, AWS CodeDeploy successfully removed all corresponding Auto Scaling lifecycle event hooks from the Amazon EC2 instances in the Auto Scaling. If the output does contain data, AWS CodeDeploy could not remove some Auto Scaling lifecycle event hooks from the Amazon EC2 instances in the Auto Scaling group.