Network.AWS.CodeDeploy.GetApplicationRevision

Description

Gets information about an application revision.

http://docs.aws.amazon.com/codedeploy/latest/APIReference/API_GetApplicationRevision.html

Request

data GetApplicationRevision

Request constructor

getApplicationRevision

Arguments

:: Text
garApplicationName

-> RevisionLocation
garRevision

-> GetApplicationRevision

GetApplicationRevision constructor.

The fields accessible through corresponding lenses are:

garApplicationName :: Text

garRevision :: RevisionLocation

Request lenses

garApplicationName :: Lens' GetApplicationRevision Text

The name of the application that corresponds to the revision.

garRevision :: Lens' GetApplicationRevision RevisionLocation

Information about the application revision to get, including the revision's type and its location.

Response

data GetApplicationRevisionResponse

Response constructor

getApplicationRevisionResponse :: GetApplicationRevisionResponse

GetApplicationRevisionResponse constructor.

The fields accessible through corresponding lenses are:

garrApplicationName :: Maybe Text

garrRevision :: Maybe RevisionLocation

garrRevisionInfo :: Maybe GenericRevisionInfo

Response lenses

garrApplicationName :: Lens' GetApplicationRevisionResponse (Maybe Text)

The name of the application that corresponds to the revision.

garrRevision :: Lens' GetApplicationRevisionResponse (Maybe RevisionLocation)

Additional information about the revision, including the revision's type and its location.

garrRevisionInfo :: Lens' GetApplicationRevisionResponse (Maybe GenericRevisionInfo)

General information about the revision.