Network.AWS.CodeDeploy.GetDeploymentConfig

Description

Gets information about a deployment configuration.

Creating a Request

getDeploymentConfig

Arguments

:: Text
gdcDeploymentConfigName

-> GetDeploymentConfig

Creates a value of GetDeploymentConfig with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdcDeploymentConfigName - The name of a deployment configuration associated with the applicable IAM user or AWS account.

data GetDeploymentConfig

Represents the input of a GetDeploymentConfig operation.

See: getDeploymentConfig smart constructor.

Request Lenses

gdcDeploymentConfigName :: Lens' GetDeploymentConfig Text

The name of a deployment configuration associated with the applicable IAM user or AWS account.

Destructuring the Response

getDeploymentConfigResponse

Arguments

:: Int
gdcrsResponseStatus

-> GetDeploymentConfigResponse

Creates a value of GetDeploymentConfigResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdcrsDeploymentConfigInfo - Information about the deployment configuration.

gdcrsResponseStatus - -- | The response status code.

data GetDeploymentConfigResponse

Represents the output of a GetDeploymentConfig operation.

See: getDeploymentConfigResponse smart constructor.

Response Lenses

gdcrsDeploymentConfigInfo :: Lens' GetDeploymentConfigResponse (Maybe DeploymentConfigInfo)

Information about the deployment configuration.

gdcrsResponseStatus :: Lens' GetDeploymentConfigResponse Int

- | The response status code.