Network.AWS.CodeDeploy.GetDeploymentInstance

Description

Gets information about an instance as part of a deployment.

Creating a Request

getDeploymentInstance

Arguments

:: Text
gdiDeploymentId

-> Text
gdiInstanceId

-> GetDeploymentInstance

Creates a value of GetDeploymentInstance with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdiDeploymentId - The unique ID of a deployment.

gdiInstanceId - The unique ID of an instance in the deployment group.

data GetDeploymentInstance

Represents the input of a GetDeploymentInstance operation.

See: getDeploymentInstance smart constructor.

Request Lenses

gdiDeploymentId :: Lens' GetDeploymentInstance Text

The unique ID of a deployment.

gdiInstanceId :: Lens' GetDeploymentInstance Text

The unique ID of an instance in the deployment group.

Destructuring the Response

getDeploymentInstanceResponse

Arguments

:: Int
gdirsResponseStatus

-> GetDeploymentInstanceResponse

Creates a value of GetDeploymentInstanceResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdirsInstanceSummary - Information about the instance.

gdirsResponseStatus - -- | The response status code.

data GetDeploymentInstanceResponse

Represents the output of a GetDeploymentInstance operation.

See: getDeploymentInstanceResponse smart constructor.

Response Lenses

gdirsInstanceSummary :: Lens' GetDeploymentInstanceResponse (Maybe InstanceSummary)

Information about the instance.

gdirsResponseStatus :: Lens' GetDeploymentInstanceResponse Int

- | The response status code.