Network.AWS.CodeDeploy.ListApplicationRevisions

Description

Lists information about revisions for an application.

http://docs.aws.amazon.com/codedeploy/latest/APIReference/API_ListApplicationRevisions.html

Request

data ListApplicationRevisions

Request constructor

listApplicationRevisions

Arguments

:: Text
larApplicationName

-> ListApplicationRevisions

ListApplicationRevisions constructor.

The fields accessible through corresponding lenses are:

larApplicationName :: Text

larDeployed :: Maybe ListStateFilterAction

larNextToken :: Maybe Text

larS3Bucket :: Maybe Text

larS3KeyPrefix :: Maybe Text

larSortBy :: Maybe ApplicationRevisionSortBy

larSortOrder :: Maybe SortOrder

Request lenses

larApplicationName :: Lens' ListApplicationRevisions Text

The name of an existing AWS CodeDeploy application associated with the applicable IAM user or AWS account.

larDeployed :: Lens' ListApplicationRevisions (Maybe ListStateFilterAction)

Whether to list revisions based on whether the revision is the target revision of an deployment group:

include: List revisions that are target revisions of a deployment group. exclude: Do not list revisions that are target revisions of a deployment group. ignore: List all revisions, regardless of whether they are target revisions of a deployment group.

larNextToken :: Lens' ListApplicationRevisions (Maybe Text)

An identifier that was returned from the previous list application revisions call, which can be used to return the next set of applications in the list.

larS3Bucket :: Lens' ListApplicationRevisions (Maybe Text)

A specific Amazon S3 bucket name to limit the search for revisions.

If set to null, then all of the user's buckets will be searched.

larS3KeyPrefix :: Lens' ListApplicationRevisions (Maybe Text)

A specific key prefix for the set of Amazon S3 objects to limit the search for revisions.

larSortBy :: Lens' ListApplicationRevisions (Maybe ApplicationRevisionSortBy)

The column name to sort the list results by:

registerTime: Sort the list results by when the revisions were registered with AWS CodeDeploy. firstUsedTime: Sort the list results by when the revisions were first used by in a deployment. lastUsedTime: Sort the list results by when the revisions were last used in a deployment. If not specified or set to null, the results will be returned in an arbitrary order.

larSortOrder :: Lens' ListApplicationRevisions (Maybe SortOrder)

The order to sort the list results by:

ascending: Sort the list of results in ascending order. descending: Sort the list of results in descending order. If not specified, the results will be sorted in ascending order.

If set to null, the results will be sorted in an arbitrary order.

Response

data ListApplicationRevisionsResponse

Response constructor

listApplicationRevisionsResponse :: ListApplicationRevisionsResponse

ListApplicationRevisionsResponse constructor.

The fields accessible through corresponding lenses are:

larrNextToken :: Maybe Text

larrRevisions :: [RevisionLocation]

Response lenses

larrNextToken :: Lens' ListApplicationRevisionsResponse (Maybe Text)

If the amount of information that is returned is significantly large, an identifier will also be returned, which can be used in a subsequent list application revisions call to return the next set of application revisions in the list.

larrRevisions :: Lens' ListApplicationRevisionsResponse [RevisionLocation]

A list of revision locations that contain the matching revisions.