Network.AWS.CodeDeploy.ListApplications

Description

Lists the applications registered with the applicable IAM user or AWS account.

http://docs.aws.amazon.com/codedeploy/latest/APIReference/API_ListApplications.html

Request

data ListApplications

Request constructor

listApplications :: ListApplications

ListApplications constructor.

The fields accessible through corresponding lenses are:

laNextToken :: Maybe Text

Request lenses

laNextToken :: Lens' ListApplications (Maybe Text)

An identifier that was returned from the previous list applications call, which can be used to return the next set of applications in the list.

Response

data ListApplicationsResponse

Response constructor

listApplicationsResponse :: ListApplicationsResponse

ListApplicationsResponse constructor.

The fields accessible through corresponding lenses are:

lar1Applications :: [Text]

lar1NextToken :: Maybe Text

Response lenses

lar1Applications :: Lens' ListApplicationsResponse [Text]

A list of application names.

lar1NextToken :: Lens' ListApplicationsResponse (Maybe Text)

If the amount of information that is returned is significantly large, an identifier will also be returned, which can be used in a subsequent list applications call to return the next set of applications in the list.