Network.AWS.CodeDeploy.ListDeploymentGroups

Description

Lists the deployment groups for an application registered with the applicable IAM user or AWS account.

http://docs.aws.amazon.com/codedeploy/latest/APIReference/API_ListDeploymentGroups.html

Request

data ListDeploymentGroups

Request constructor

listDeploymentGroups

Arguments

:: Text
ldgApplicationName

-> ListDeploymentGroups

ListDeploymentGroups constructor.

The fields accessible through corresponding lenses are:

ldgApplicationName :: Text

ldgNextToken :: Maybe Text

Request lenses

ldgApplicationName :: Lens' ListDeploymentGroups Text

The name of an existing AWS CodeDeploy application associated with the applicable IAM user or AWS account.

ldgNextToken :: Lens' ListDeploymentGroups (Maybe Text)

An identifier that was returned from the previous list deployment groups call, which can be used to return the next set of deployment groups in the list.

Response

data ListDeploymentGroupsResponse

Response constructor

listDeploymentGroupsResponse :: ListDeploymentGroupsResponse

ListDeploymentGroupsResponse constructor.

The fields accessible through corresponding lenses are:

ldgrApplicationName :: Maybe Text

ldgrDeploymentGroups :: [Text]

ldgrNextToken :: Maybe Text

Response lenses

ldgrApplicationName :: Lens' ListDeploymentGroupsResponse (Maybe Text)

The application name.

ldgrDeploymentGroups :: Lens' ListDeploymentGroupsResponse [Text]

A list of corresponding deployment group names.

ldgrNextToken :: Lens' ListDeploymentGroupsResponse (Maybe Text)

If the amount of information that is returned is significantly large, an identifier will also be returned, which can be used in a subsequent list deployment groups call to return the next set of deployment groups in the list.