Network.AWS.CodeDeploy.ListDeploymentInstances

Description

Lists the instances for a deployment associated with the applicable IAM user or AWS account.

http://docs.aws.amazon.com/codedeploy/latest/APIReference/API_ListDeploymentInstances.html

Request

data ListDeploymentInstances

Request constructor

listDeploymentInstances

Arguments

:: Text
ldiDeploymentId

-> ListDeploymentInstances

ListDeploymentInstances constructor.

The fields accessible through corresponding lenses are:

ldiDeploymentId :: Text

ldiInstanceStatusFilter :: [InstanceStatus]

ldiNextToken :: Maybe Text

Request lenses

ldiDeploymentId :: Lens' ListDeploymentInstances Text

The unique ID of a deployment.

ldiInstanceStatusFilter :: Lens' ListDeploymentInstances [InstanceStatus]

A subset of instances to list, by status:

Pending: Include in the resulting list those instances with pending deployments. InProgress: Include in the resulting list those instances with in-progress deployments. Succeeded: Include in the resulting list those instances with succeeded deployments. Failed: Include in the resulting list those instances with failed deployments. Skipped: Include in the resulting list those instances with skipped deployments. Unknown: Include in the resulting list those instances with deployments in an unknown state.

ldiNextToken :: Lens' ListDeploymentInstances (Maybe Text)

An identifier that was returned from the previous list deployment instances call, which can be used to return the next set of deployment instances in the list.

Response

data ListDeploymentInstancesResponse

Response constructor

listDeploymentInstancesResponse :: ListDeploymentInstancesResponse

ListDeploymentInstancesResponse constructor.

The fields accessible through corresponding lenses are:

ldirInstancesList :: [Text]

ldirNextToken :: Maybe Text

Response lenses

ldirInstancesList :: Lens' ListDeploymentInstancesResponse [Text]

A list of instances IDs.

ldirNextToken :: Lens' ListDeploymentInstancesResponse (Maybe Text)

If the amount of information that is returned is significantly large, an identifier will also be returned, which can be used in a subsequent list deployment instances call to return the next set of deployment instances in the list.