Network.AWS.CodeDeploy.ListDeployments

Description

Lists the deployments within a deployment group for an application registered with the applicable IAM user or AWS account.

http://docs.aws.amazon.com/codedeploy/latest/APIReference/API_ListDeployments.html

Request

data ListDeployments

Request constructor

listDeployments :: ListDeployments

ListDeployments constructor.

The fields accessible through corresponding lenses are:

ldApplicationName :: Maybe Text

ldCreateTimeRange :: Maybe TimeRange

ldDeploymentGroupName :: Maybe Text

ldIncludeOnlyStatuses :: [DeploymentStatus]

ldNextToken :: Maybe Text

Request lenses

ldApplicationName :: Lens' ListDeployments (Maybe Text)

The name of an existing AWS CodeDeploy application associated with the applicable IAM user or AWS account.

ldCreateTimeRange :: Lens' ListDeployments (Maybe TimeRange)

A deployment creation start- and end-time range for returning a subset of the list of deployments.

ldDeploymentGroupName :: Lens' ListDeployments (Maybe Text)

The name of an existing deployment group for the specified application.

ldIncludeOnlyStatuses :: Lens' ListDeployments [DeploymentStatus]

A subset of deployments to list, by status:

Created: Include in the resulting list created deployments. Queued: Include in the resulting list queued deployments. In Progress: Include in the resulting list in-progress deployments. Succeeded: Include in the resulting list succeeded deployments. Failed: Include in the resulting list failed deployments. Aborted: Include in the resulting list aborted deployments.

ldNextToken :: Lens' ListDeployments (Maybe Text)

An identifier that was returned from the previous list deployments call, which can be used to return the next set of deployments in the list.

Response

data ListDeploymentsResponse

Response constructor

listDeploymentsResponse :: ListDeploymentsResponse

ListDeploymentsResponse constructor.

The fields accessible through corresponding lenses are:

ldrDeployments :: [Text]

ldrNextToken :: Maybe Text

Response lenses

ldrDeployments :: Lens' ListDeploymentsResponse [Text]

A list of deployment IDs.

ldrNextToken :: Lens' ListDeploymentsResponse (Maybe Text)

If the amount of information that is returned is significantly large, an identifier will also be returned, which can be used in a subsequent list deployments call to return the next set of deployments in the list.