Network.AWS.CodeDeploy.ListOnPremisesInstances

Description

Gets a list of one or more on-premises instance names.

Unless otherwise specified, both registered and deregistered on-premises instance names will be listed. To list only registered or deregistered on-premises instance names, use the registration status parameter.

http://docs.aws.amazon.com/codedeploy/latest/APIReference/API_ListOnPremisesInstances.html

Request

data ListOnPremisesInstances

Request constructor

listOnPremisesInstances :: ListOnPremisesInstances

ListOnPremisesInstances constructor.

The fields accessible through corresponding lenses are:

lopiNextToken :: Maybe Text

lopiRegistrationStatus :: Maybe RegistrationStatus

lopiTagFilters :: [TagFilter]

Request lenses

lopiNextToken :: Lens' ListOnPremisesInstances (Maybe Text)

An identifier that was returned from the previous list on-premises instances call, which can be used to return the next set of on-premises instances in the list.

lopiRegistrationStatus :: Lens' ListOnPremisesInstances (Maybe RegistrationStatus)

The on-premises instances registration status:

Deregistered: Include in the resulting list deregistered on-premises instances. Registered: Include in the resulting list registered on-premises instances.

lopiTagFilters :: Lens' ListOnPremisesInstances [TagFilter]

The on-premises instance tags that will be used to restrict the corresponding on-premises instance names that are returned.

Response

data ListOnPremisesInstancesResponse

Response constructor

listOnPremisesInstancesResponse :: ListOnPremisesInstancesResponse

ListOnPremisesInstancesResponse constructor.

The fields accessible through corresponding lenses are:

lopirInstanceNames :: [Text]

lopirNextToken :: Maybe Text

Response lenses

lopirInstanceNames :: Lens' ListOnPremisesInstancesResponse [Text]

The list of matching on-premises instance names.

lopirNextToken :: Lens' ListOnPremisesInstancesResponse (Maybe Text)

If the amount of information that is returned is significantly large, an identifier will also be returned, which can be used in a subsequent list on-premises instances call to return the next set of on-premises instances in the list.