Network.AWS.CodeDeploy.StopDeployment

Description

Attempts to stop an ongoing deployment.

Creating a Request

stopDeployment

Arguments

:: Text
sdDeploymentId

-> StopDeployment

Creates a value of StopDeployment with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sdAutoRollbackEnabled - Indicates, when a deployment is stopped, whether instances that have been updated should be rolled back to the previous version of the application revision.

sdDeploymentId - The unique ID of a deployment.

data StopDeployment

Represents the input of a StopDeployment operation.

See: stopDeployment smart constructor.

Request Lenses

sdAutoRollbackEnabled :: Lens' StopDeployment (Maybe Bool)

Indicates, when a deployment is stopped, whether instances that have been updated should be rolled back to the previous version of the application revision.

sdDeploymentId :: Lens' StopDeployment Text

The unique ID of a deployment.

Destructuring the Response

stopDeploymentResponse

Arguments

:: Int
sdrsResponseStatus

-> StopDeploymentResponse

Creates a value of StopDeploymentResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sdrsStatus - The status of the stop deployment operation: * Pending: The stop operation is pending. * Succeeded: The stop operation was successful.

sdrsStatusMessage - An accompanying status message.

sdrsResponseStatus - -- | The response status code.

data StopDeploymentResponse

Represents the output of a StopDeployment operation.

See: stopDeploymentResponse smart constructor.

Response Lenses

sdrsStatus :: Lens' StopDeploymentResponse (Maybe StopStatus)

The status of the stop deployment operation: * Pending: The stop operation is pending. * Succeeded: The stop operation was successful.

sdrsStatusMessage :: Lens' StopDeploymentResponse (Maybe Text)

An accompanying status message.

sdrsResponseStatus :: Lens' StopDeploymentResponse Int

- | The response status code.