Network.AWS.CodeDeploy.Types

Service

data CodeDeploy

Version 2014-10-06 of the Amazon CodeDeploy service.

Error

data JSONError :: *

GenericRevisionInfo

data GenericRevisionInfo

genericRevisionInfo :: GenericRevisionInfo

GenericRevisionInfo constructor.

The fields accessible through corresponding lenses are:

griDeploymentGroups :: [Text]

griDescription :: Maybe Text

griFirstUsedTime :: Maybe UTCTime

griLastUsedTime :: Maybe UTCTime

griRegisterTime :: Maybe UTCTime

griDeploymentGroups :: Lens' GenericRevisionInfo [Text]

A list of deployment groups that use this revision.

griDescription :: Lens' GenericRevisionInfo (Maybe Text)

A comment about the revision.

griFirstUsedTime :: Lens' GenericRevisionInfo (Maybe UTCTime)

When the revision was first used by AWS CodeDeploy.

griLastUsedTime :: Lens' GenericRevisionInfo (Maybe UTCTime)

When the revision was last used by AWS CodeDeploy.

griRegisterTime :: Lens' GenericRevisionInfo (Maybe UTCTime)

When the revision was registered with AWS CodeDeploy.

ApplicationInfo

data ApplicationInfo

applicationInfo :: ApplicationInfo

ApplicationInfo constructor.

The fields accessible through corresponding lenses are:

aiApplicationId :: Maybe Text

aiApplicationName :: Maybe Text

aiCreateTime :: Maybe UTCTime

aiLinkedToGitHub :: Maybe Bool

aiApplicationId :: Lens' ApplicationInfo (Maybe Text)

The application ID.

aiApplicationName :: Lens' ApplicationInfo (Maybe Text)

The application name.

aiCreateTime :: Lens' ApplicationInfo (Maybe UTCTime)

The time that the application was created.

aiLinkedToGitHub :: Lens' ApplicationInfo (Maybe Bool)

True if the user has authenticated with GitHub for the specified application; otherwise, false.

BundleType

data BundleType

Constructors

Tar
tar

Tgz
tgz

Zip
zip

Tag

data Tag

tag :: Tag

Tag constructor.

The fields accessible through corresponding lenses are:

tagKey :: Maybe Text

tagValue :: Maybe Text

tagKey :: Lens' Tag (Maybe Text)

The tag's key.

tagValue :: Lens' Tag (Maybe Text)

The tag's value.

TagFilterType

data TagFilterType

Constructors

KeyAndValue
KEY_AND_VALUE

KeyOnly
KEY_ONLY

ValueOnly
VALUE_ONLY

TimeRange

data TimeRange

timeRange :: TimeRange

TimeRange constructor.

The fields accessible through corresponding lenses are:

trEnd :: Maybe UTCTime

trStart :: Maybe UTCTime

trEnd :: Lens' TimeRange (Maybe UTCTime)

The time range's end time.

Specify null to leave the time range's end time open-ended.

trStart :: Lens' TimeRange (Maybe UTCTime)

The time range's start time.

Specify null to leave the time range's start time open-ended.

DeploymentCreator

data DeploymentCreator

Constructors

Autoscaling
autoscaling

User
user

InstanceSummary

data InstanceSummary

instanceSummary :: InstanceSummary

InstanceSummary constructor.

The fields accessible through corresponding lenses are:

isDeploymentId :: Maybe Text

isInstanceId :: Maybe Text

isLastUpdatedAt :: Maybe UTCTime

isLifecycleEvents :: [LifecycleEvent]

isStatus :: Maybe InstanceStatus

isDeploymentId :: Lens' InstanceSummary (Maybe Text)

The deployment ID.

isInstanceId :: Lens' InstanceSummary (Maybe Text)

The instance ID.

isLastUpdatedAt :: Lens' InstanceSummary (Maybe UTCTime)

A timestamp indicating when the instance information was last updated.

isLifecycleEvents :: Lens' InstanceSummary [LifecycleEvent]

A list of lifecycle events for this instance.

isStatus :: Lens' InstanceSummary (Maybe InstanceStatus)

The deployment status for this instance:

Pending: The deployment is pending for this instance. In Progress: The deployment is in progress for this instance. Succeeded: The deployment has succeeded for this instance. Failed: The deployment has failed for this instance. Skipped: The deployment has been skipped for this instance. Unknown: The deployment status is unknown for this instance.

AutoScalingGroup

data AutoScalingGroup

autoScalingGroup :: AutoScalingGroup

AutoScalingGroup constructor.

The fields accessible through corresponding lenses are:

asgHook :: Maybe Text

asgName :: Maybe Text

asgHook :: Lens' AutoScalingGroup (Maybe Text)

An Auto Scaling lifecycle event hook name.

asgName :: Lens' AutoScalingGroup (Maybe Text)

The Auto Scaling group name.

DeploymentGroupInfo

data DeploymentGroupInfo

deploymentGroupInfo :: DeploymentGroupInfo

DeploymentGroupInfo constructor.

The fields accessible through corresponding lenses are:

dgiApplicationName :: Maybe Text

dgiAutoScalingGroups :: [AutoScalingGroup]

dgiDeploymentConfigName :: Maybe Text

dgiDeploymentGroupId :: Maybe Text

dgiDeploymentGroupName :: Maybe Text

dgiEc2TagFilters :: [EC2TagFilter]

dgiOnPremisesInstanceTagFilters :: [TagFilter]

dgiServiceRoleArn :: Maybe Text

dgiTargetRevision :: Maybe RevisionLocation

dgiApplicationName :: Lens' DeploymentGroupInfo (Maybe Text)

The application name.

dgiAutoScalingGroups :: Lens' DeploymentGroupInfo [AutoScalingGroup]

A list of associated Auto Scaling groups.

dgiDeploymentConfigName :: Lens' DeploymentGroupInfo (Maybe Text)

The deployment configuration name.

dgiDeploymentGroupId :: Lens' DeploymentGroupInfo (Maybe Text)

The deployment group ID.

dgiDeploymentGroupName :: Lens' DeploymentGroupInfo (Maybe Text)

The deployment group name.

dgiEc2TagFilters :: Lens' DeploymentGroupInfo [EC2TagFilter]

The Amazon EC2 tags to filter on.

dgiOnPremisesInstanceTagFilters :: Lens' DeploymentGroupInfo [TagFilter]

The on-premises instance tags to filter on.

dgiServiceRoleArn :: Lens' DeploymentGroupInfo (Maybe Text)

A service role ARN.

dgiTargetRevision :: Lens' DeploymentGroupInfo (Maybe RevisionLocation)

Information about the deployment group's target revision, including the revision's type and its location.

ApplicationRevisionSortBy

data ApplicationRevisionSortBy

Constructors

FirstUsedTime
firstUsedTime

LastUsedTime
lastUsedTime

RegisterTime
registerTime

MinimumHealthyHosts

data MinimumHealthyHosts

minimumHealthyHosts :: MinimumHealthyHosts

MinimumHealthyHosts constructor.

The fields accessible through corresponding lenses are:

mhhType :: Maybe MinimumHealthyHostsType

mhhValue :: Maybe Int

mhhType :: Lens' MinimumHealthyHosts (Maybe MinimumHealthyHostsType)

The minimum healthy instances type:

HOST_COUNT: The minimum number of healthy instances, as an absolute value. FLEET_PERCENT: The minimum number of healthy instances, as a percentage of the total number of instances in the deployment. For example, for 9 instances, if a HOST_COUNT of 6 is specified, deploy to up to 3 instances at a time. The deployment succeeds if 6 or more instances are successfully deployed to; otherwise, the deployment fails. If a FLEET_PERCENT of 40 is specified, deploy to up to 5 instances at a time. The deployment succeeds if 4 or more instances are successfully deployed to; otherwise, the deployment fails.

In a call to the get deployment configuration operation, CodeDeployDefault.OneAtATime will return a minimum healthy instances type of MOST_CONCURRENCY and a value of 1. This means a deployment to only one instances at a time. (You cannot set the type to MOST_CONCURRENCY, only to HOST_COUNT or FLEET_PERCENT.)

mhhValue :: Lens' MinimumHealthyHosts (Maybe Int)

The minimum healthy instances value.

ListStateFilterAction

data ListStateFilterAction

Constructors

Exclude
exclude

Ignore
ignore

Include
include

LifecycleErrorCode

data LifecycleErrorCode

Constructors

ScriptFailed
ScriptFailed

ScriptMissing
ScriptMissing

ScriptNotExecutable
ScriptNotExecutable

ScriptTimedOut
ScriptTimedOut

Success
Success

UnknownError
UnknownError

RevisionLocation

data RevisionLocation

revisionLocation :: RevisionLocation

RevisionLocation constructor.

The fields accessible through corresponding lenses are:

rlGitHubLocation :: Maybe GitHubLocation

rlRevisionType :: Maybe RevisionLocationType

rlS3Location :: Maybe S3Location

rlGitHubLocation :: Lens' RevisionLocation (Maybe GitHubLocation)

rlRevisionType :: Lens' RevisionLocation (Maybe RevisionLocationType)

The application revision's type:

S3: An application revision stored in Amazon S3. GitHub: An application revision stored in GitHub.

rlS3Location :: Lens' RevisionLocation (Maybe S3Location)

LifecycleEventStatus

data LifecycleEventStatus

Constructors

Failed
Failed

InProgress
InProgress

Pending
Pending

Skipped
Skipped

Succeeded
Succeeded

Unknown
Unknown

EC2TagFilter

data EC2TagFilter

ec2TagFilter :: EC2TagFilter

EC2TagFilter constructor.

The fields accessible through corresponding lenses are:

ectfKey :: Maybe Text

ectfType :: Maybe EC2TagFilterType

ectfValue :: Maybe Text

ectfKey :: Lens' EC2TagFilter (Maybe Text)

The tag filter key.

ectfType :: Lens' EC2TagFilter (Maybe EC2TagFilterType)

The tag filter type:

KEY_ONLY: Key only. VALUE_ONLY: Value only. KEY_AND_VALUE: Key and value.

ectfValue :: Lens' EC2TagFilter (Maybe Text)

The tag filter value.

Diagnostics

data Diagnostics

diagnostics :: Diagnostics

Diagnostics constructor.

The fields accessible through corresponding lenses are:

dErrorCode :: Maybe LifecycleErrorCode

dLogTail :: Maybe Text

dMessage :: Maybe Text

dScriptName :: Maybe Text

dErrorCode :: Lens' Diagnostics (Maybe LifecycleErrorCode)

The associated error code:

Success: The specified script ran. ScriptMissing: The specified script was not found in the specified location. ScriptNotExecutable: The specified script is not a recognized executable file type. ScriptTimedOut: The specified script did not finish running in the specified time period. ScriptFailed: The specified script failed to run as expected. UnknownError: The specified script did not run for an unknown reason.

dLogTail :: Lens' Diagnostics (Maybe Text)

The last portion of the associated diagnostic log.

dMessage :: Lens' Diagnostics (Maybe Text)

The message associated with the error.

dScriptName :: Lens' Diagnostics (Maybe Text)

The name of the script.

StopStatus

data StopStatus

Constructors

SSPending
Pending

SSSucceeded
Succeeded

ErrorInformation

data ErrorInformation

errorInformation :: ErrorInformation

ErrorInformation constructor.

The fields accessible through corresponding lenses are:

eiCode :: Maybe ErrorCode

eiMessage :: Maybe Text

eiCode :: Lens' ErrorInformation (Maybe ErrorCode)

The error code:

APPLICATION_MISSING: The application was missing. Note that this error code will most likely be raised if the application is deleted after the deployment is created but before it starts. DEPLOYMENT_GROUP_MISSING: The deployment group was missing. Note that this error code will most likely be raised if the deployment group is deleted after the deployment is created but before it starts. HEALTH_CONSTRAINTS: The deployment failed on too many instances to be able to successfully deploy within the specified instance health constraints. HEALTH_CONSTRAINTS_INVALID: The revision can never successfully deploy within the instance health constraints as specified. IAM_ROLE_MISSING: The service role cannot be accessed. IAM_ROLE_PERMISSIONS: The service role does not have the correct permissions. INTERNAL_ERROR: There was an internal error. NO_EC2_SUBSCRIPTION: The calling account is not subscribed to the Amazon EC2 service. NO_INSTANCES: No instances were specified, or no instances can be found. OVER_MAX_INSTANCES: The maximum number of instances was exceeded. THROTTLED: The operation was throttled because the calling account exceeded the throttling limits of one or more AWS services. TIMEOUT: The deployment has timed out. REVISION_MISSING: The revision ID was missing. Note that this error code will most likely be raised if the revision is deleted after the deployment is created but before it starts.

eiMessage :: Lens' ErrorInformation (Maybe Text)

An accompanying error message.

SortOrder

data SortOrder

Constructors

Ascending
ascending

Descending
descending

InstanceInfo

data InstanceInfo

instanceInfo :: InstanceInfo

InstanceInfo constructor.

The fields accessible through corresponding lenses are:

iiDeregisterTime :: Maybe UTCTime

iiIamUserArn :: Maybe Text

iiInstanceArn :: Maybe Text

iiInstanceName :: Maybe Text

iiRegisterTime :: Maybe UTCTime

iiTags :: [Tag]

iiDeregisterTime :: Lens' InstanceInfo (Maybe UTCTime)

If the on-premises instance was deregistered, the time that the on-premises instance was deregistered.

iiIamUserArn :: Lens' InstanceInfo (Maybe Text)

The IAM user ARN associated with the on-premises instance.

iiInstanceArn :: Lens' InstanceInfo (Maybe Text)

The ARN of the on-premises instance.

iiInstanceName :: Lens' InstanceInfo (Maybe Text)

The name of the on-premises instance.

iiRegisterTime :: Lens' InstanceInfo (Maybe UTCTime)

The time that the on-premises instance was registered.

iiTags :: Lens' InstanceInfo [Tag]

The tags that are currently associated with the on-premises instance.

DeploymentInfo

data DeploymentInfo

deploymentInfo :: DeploymentInfo

DeploymentInfo constructor.

The fields accessible through corresponding lenses are:

diApplicationName :: Maybe Text

diCompleteTime :: Maybe UTCTime

diCreateTime :: Maybe UTCTime

diCreator :: Maybe DeploymentCreator

diDeploymentConfigName :: Maybe Text

diDeploymentGroupName :: Maybe Text

diDeploymentId :: Maybe Text

diDeploymentOverview :: Maybe DeploymentOverview

diDescription :: Maybe Text

diErrorInformation :: Maybe ErrorInformation

diIgnoreApplicationStopFailures :: Maybe Bool

diRevision :: Maybe RevisionLocation

diStartTime :: Maybe UTCTime

diStatus :: Maybe DeploymentStatus

diApplicationName :: Lens' DeploymentInfo (Maybe Text)

The application name.

diCompleteTime :: Lens' DeploymentInfo (Maybe UTCTime)

A timestamp indicating when the deployment was completed.

diCreateTime :: Lens' DeploymentInfo (Maybe UTCTime)

A timestamp indicating when the deployment was created.

diCreator :: Lens' DeploymentInfo (Maybe DeploymentCreator)

How the deployment was created:

user: A user created the deployment. autoscaling: Auto Scaling created the deployment.

diDeploymentConfigName :: Lens' DeploymentInfo (Maybe Text)

The deployment configuration name.

diDeploymentGroupName :: Lens' DeploymentInfo (Maybe Text)

The deployment group name.

diDeploymentId :: Lens' DeploymentInfo (Maybe Text)

The deployment ID.

diDeploymentOverview :: Lens' DeploymentInfo (Maybe DeploymentOverview)

A summary of the deployment status of the instances in the deployment.

diDescription :: Lens' DeploymentInfo (Maybe Text)

A comment about the deployment.

diErrorInformation :: Lens' DeploymentInfo (Maybe ErrorInformation)

Information about any error associated with this deployment.

diIgnoreApplicationStopFailures :: Lens' DeploymentInfo (Maybe Bool)

If true, then if the deployment causes the ApplicationStop deployment lifecycle event to fail to a specific instance, the deployment will not be considered to have failed to that instance at that point and will continue on to the BeforeInstall deployment lifecycle event.

If false or not specified, then if the deployment causes the ApplicationStop deployment lifecycle event to fail to a specific instance, the deployment will stop to that instance, and the deployment to that instance will be considered to have failed.

diRevision :: Lens' DeploymentInfo (Maybe RevisionLocation)

Information about the location of application artifacts that are stored and the service to retrieve them from.

diStartTime :: Lens' DeploymentInfo (Maybe UTCTime)

A timestamp indicating when the deployment began deploying to the deployment group.

Note that in some cases, the reported value of the start time may be later than the complete time. This is due to differences in the clock settings of various back-end servers that participate in the overall deployment process.

diStatus :: Lens' DeploymentInfo (Maybe DeploymentStatus)

The current state of the deployment as a whole.

TagFilter

data TagFilter

tagFilter :: TagFilter

TagFilter constructor.

The fields accessible through corresponding lenses are:

tfKey :: Maybe Text

tfType :: Maybe TagFilterType

tfValue :: Maybe Text

tfKey :: Lens' TagFilter (Maybe Text)

The on-premises instance tag filter key.

tfType :: Lens' TagFilter (Maybe TagFilterType)

The on-premises instance tag filter type:

KEY_ONLY: Key only. VALUE_ONLY: Value only. KEY_AND_VALUE: Key and value.

tfValue :: Lens' TagFilter (Maybe Text)

The on-premises instance tag filter value.

LifecycleEvent

data LifecycleEvent

lifecycleEvent :: LifecycleEvent

LifecycleEvent constructor.

The fields accessible through corresponding lenses are:

leDiagnostics :: Maybe Diagnostics

leEndTime :: Maybe UTCTime

leLifecycleEventName :: Maybe Text

leStartTime :: Maybe UTCTime

leStatus :: Maybe LifecycleEventStatus

leDiagnostics :: Lens' LifecycleEvent (Maybe Diagnostics)

Diagnostic information about the deployment lifecycle event.

leEndTime :: Lens' LifecycleEvent (Maybe UTCTime)

A timestamp indicating when the deployment lifecycle event ended.

leLifecycleEventName :: Lens' LifecycleEvent (Maybe Text)

The deployment lifecycle event name, such as ApplicationStop, BeforeInstall, AfterInstall, ApplicationStart, or ValidateService.

leStartTime :: Lens' LifecycleEvent (Maybe UTCTime)

A timestamp indicating when the deployment lifecycle event started.

leStatus :: Lens' LifecycleEvent (Maybe LifecycleEventStatus)

The deployment lifecycle event status:

Pending: The deployment lifecycle event is pending. InProgress: The deployment lifecycle event is in progress. Succeeded: The deployment lifecycle event has succeeded. Failed: The deployment lifecycle event has failed. Skipped: The deployment lifecycle event has been skipped. Unknown: The deployment lifecycle event is unknown.

DeploymentOverview

data DeploymentOverview

deploymentOverview :: DeploymentOverview

DeploymentOverview constructor.

The fields accessible through corresponding lenses are:

doFailed :: Maybe Integer

doInProgress :: Maybe Integer

doPending :: Maybe Integer

doSkipped :: Maybe Integer

doSucceeded :: Maybe Integer

doFailed :: Lens' DeploymentOverview (Maybe Integer)

The number of instances that have failed in the deployment.

doInProgress :: Lens' DeploymentOverview (Maybe Integer)

The number of instances that are in progress in the deployment.

doPending :: Lens' DeploymentOverview (Maybe Integer)

The number of instances that are pending in the deployment.

doSkipped :: Lens' DeploymentOverview (Maybe Integer)

The number of instances that have been skipped in the deployment.

doSucceeded :: Lens' DeploymentOverview (Maybe Integer)

The number of instances that have succeeded in the deployment.

ErrorCode

data ErrorCode

Constructors

ApplicationMissing
APPLICATION_MISSING

DeploymentGroupMissing
DEPLOYMENT_GROUP_MISSING

HealthConstraints
HEALTH_CONSTRAINTS

HealthConstraintsInvalid
HEALTH_CONSTRAINTS_INVALID

IamRoleMissing
IAM_ROLE_MISSING

IamRolePermissions
IAM_ROLE_PERMISSIONS

InternalError
INTERNAL_ERROR

NOEC2SUBSCRIPTION
NO_EC2_SUBSCRIPTION

NoInstances
NO_INSTANCES

OverMaxInstances
OVER_MAX_INSTANCES

RevisionMissing
REVISION_MISSING

Throttled
THROTTLED

Timeout
TIMEOUT

DeploymentConfigInfo

data DeploymentConfigInfo

deploymentConfigInfo :: DeploymentConfigInfo

DeploymentConfigInfo constructor.

The fields accessible through corresponding lenses are:

dciCreateTime :: Maybe UTCTime

dciDeploymentConfigId :: Maybe Text

dciDeploymentConfigName :: Maybe Text

dciMinimumHealthyHosts :: Maybe MinimumHealthyHosts

dciCreateTime :: Lens' DeploymentConfigInfo (Maybe UTCTime)

The time that the deployment configuration was created.

dciDeploymentConfigId :: Lens' DeploymentConfigInfo (Maybe Text)

The deployment configuration ID.

dciDeploymentConfigName :: Lens' DeploymentConfigInfo (Maybe Text)

The deployment configuration name.

dciMinimumHealthyHosts :: Lens' DeploymentConfigInfo (Maybe MinimumHealthyHosts)

Information about the number or percentage of minimum healthy instances.

InstanceStatus

data InstanceStatus

Constructors

ISFailed
Failed

ISInProgress
InProgress

ISPending
Pending

ISSkipped
Skipped

ISSucceeded
Succeeded

ISUnknown
Unknown

DeploymentStatus

data DeploymentStatus

Constructors

DSCreated
Created

DSFailed
Failed

DSInProgress
InProgress

DSQueued
Queued

DSStopped
Stopped

DSSucceeded
Succeeded

RegistrationStatus

data RegistrationStatus

Constructors

Deregistered
Deregistered

Registered
Registered

S3Location

data S3Location

s3Location :: S3Location

S3Location constructor.

The fields accessible through corresponding lenses are:

slBucket :: Maybe Text

slBundleType :: Maybe BundleType

slETag :: Maybe Text

slKey :: Maybe Text

slVersion :: Maybe Text

slBucket :: Lens' S3Location (Maybe Text)

The name of the Amazon S3 bucket where the application revision is stored.

slBundleType :: Lens' S3Location (Maybe BundleType)

The file type of the application revision. Must be one of the following:

tar: A tar archive file. tgz: A compressed tar archive file. zip: A zip archive file.

slETag :: Lens' S3Location (Maybe Text)

The ETag of the Amazon S3 object that represents the bundled artifacts for the application revision.

If the ETag is not specified as an input parameter, ETag validation of the object will be skipped.

slKey :: Lens' S3Location (Maybe Text)

The name of the Amazon S3 object that represents the bundled artifacts for the application revision.

slVersion :: Lens' S3Location (Maybe Text)

A specific version of the Amazon S3 object that represents the bundled artifacts for the application revision.

If the version is not specified, the system will use the most recent version by default.

MinimumHealthyHostsType

data MinimumHealthyHostsType

Constructors

FleetPercent
FLEET_PERCENT

HostCount
HOST_COUNT

GitHubLocation

data GitHubLocation

gitHubLocation :: GitHubLocation

GitHubLocation constructor.

The fields accessible through corresponding lenses are:

ghlCommitId :: Maybe Text

ghlRepository :: Maybe Text

ghlCommitId :: Lens' GitHubLocation (Maybe Text)

The SHA1 commit ID of the GitHub commit that references the that represents the bundled artifacts for the application revision.

ghlRepository :: Lens' GitHubLocation (Maybe Text)

The GitHub account and repository pair that stores a reference to the commit that represents the bundled artifacts for the application revision.

Specified as account/repository.

RevisionLocationType

data RevisionLocationType

Constructors

GitHub
GitHub

S3
S3

EC2TagFilterType

data EC2TagFilterType

Constructors

ECTFTKeyAndValue
KEY_AND_VALUE

ECTFTKeyOnly
KEY_ONLY

ECTFTValueOnly
VALUE_ONLY