Network.AWS.CodeDeploy.UpdateDeploymentGroup

Description

Changes information about an existing deployment group.

http://docs.aws.amazon.com/codedeploy/latest/APIReference/API_UpdateDeploymentGroup.html

Request

data UpdateDeploymentGroup

Request constructor

updateDeploymentGroup

Arguments

:: Text
udgApplicationName

-> Text
udgCurrentDeploymentGroupName

-> UpdateDeploymentGroup

UpdateDeploymentGroup constructor.

The fields accessible through corresponding lenses are:

udgApplicationName :: Text

udgAutoScalingGroups :: [Text]

udgCurrentDeploymentGroupName :: Text

udgDeploymentConfigName :: Maybe Text

udgEc2TagFilters :: [EC2TagFilter]

udgNewDeploymentGroupName :: Maybe Text

udgOnPremisesInstanceTagFilters :: [TagFilter]

udgServiceRoleArn :: Maybe Text

Request lenses

udgApplicationName :: Lens' UpdateDeploymentGroup Text

The application name corresponding to the deployment group to update.

udgAutoScalingGroups :: Lens' UpdateDeploymentGroup [Text]

The replacement list of Auto Scaling groups to be included in the deployment group, if you want to change them.

udgCurrentDeploymentGroupName :: Lens' UpdateDeploymentGroup Text

The current name of the existing deployment group.

udgDeploymentConfigName :: Lens' UpdateDeploymentGroup (Maybe Text)

The replacement deployment configuration name to use, if you want to change it.

udgEc2TagFilters :: Lens' UpdateDeploymentGroup [EC2TagFilter]

The replacement set of Amazon EC2 tags to filter on, if you want to change them.

udgNewDeploymentGroupName :: Lens' UpdateDeploymentGroup (Maybe Text)

The new name of the deployment group, if you want to change it.

udgOnPremisesInstanceTagFilters :: Lens' UpdateDeploymentGroup [TagFilter]

The replacement set of on-premises instance tags for filter on, if you want to change them.

udgServiceRoleArn :: Lens' UpdateDeploymentGroup (Maybe Text)

A replacement service role's ARN, if you want to change it.

Response

data UpdateDeploymentGroupResponse

Response constructor

updateDeploymentGroupResponse :: UpdateDeploymentGroupResponse

UpdateDeploymentGroupResponse constructor.

The fields accessible through corresponding lenses are:

udgrHooksNotCleanedUp :: [AutoScalingGroup]

Response lenses

udgrHooksNotCleanedUp :: Lens' UpdateDeploymentGroupResponse [AutoScalingGroup]

If the output contains no data, and the corresponding deployment group contained at least one Auto Scaling group, AWS CodeDeploy successfully removed all corresponding Auto Scaling lifecycle event hooks from the AWS account. If the output does contain data, AWS CodeDeploy could not remove some Auto Scaling lifecycle event hooks from the AWS account.