Network.AWS.CodeDeploy.CreateDeploymentConfig

Description

Creates a deployment configuration.

Creating a Request

createDeploymentConfig

Arguments

:: Text
cdcDeploymentConfigName

-> MinimumHealthyHosts
cdcMinimumHealthyHosts

-> CreateDeploymentConfig

Creates a value of CreateDeploymentConfig with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdcDeploymentConfigName - The name of the deployment configuration to create.

cdcMinimumHealthyHosts - The minimum number of healthy instances that should be available at any time during the deployment. There are two parameters expected in the input: type and value. The type parameter takes either of the following values: * HOST_COUNT: The value parameter represents the minimum number of healthy instances as an absolute value. * FLEET_PERCENT: The value parameter represents the minimum number of healthy instances as a percentage of the total number of instances in the deployment. If you specify FLEET_PERCENT, at the start of the deployment, AWS CodeDeploy converts the percentage to the equivalent number of instance and rounds up fractional instances. The value parameter takes an integer. For example, to set a minimum of 95% healthy instance, specify a type of FLEET_PERCENT and a value of 95.

data CreateDeploymentConfig

Represents the input of a CreateDeploymentConfig operation.

See: createDeploymentConfig smart constructor.

Request Lenses

cdcDeploymentConfigName :: Lens' CreateDeploymentConfig Text

The name of the deployment configuration to create.

cdcMinimumHealthyHosts :: Lens' CreateDeploymentConfig MinimumHealthyHosts

The minimum number of healthy instances that should be available at any time during the deployment. There are two parameters expected in the input: type and value. The type parameter takes either of the following values: * HOST_COUNT: The value parameter represents the minimum number of healthy instances as an absolute value. * FLEET_PERCENT: The value parameter represents the minimum number of healthy instances as a percentage of the total number of instances in the deployment. If you specify FLEET_PERCENT, at the start of the deployment, AWS CodeDeploy converts the percentage to the equivalent number of instance and rounds up fractional instances. The value parameter takes an integer. For example, to set a minimum of 95% healthy instance, specify a type of FLEET_PERCENT and a value of 95.

Destructuring the Response

createDeploymentConfigResponse

Arguments

:: Int
cdcrsResponseStatus

-> CreateDeploymentConfigResponse

Creates a value of CreateDeploymentConfigResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdcrsDeploymentConfigId - A unique deployment configuration ID.

cdcrsResponseStatus - -- | The response status code.

data CreateDeploymentConfigResponse

Represents the output of a CreateDeploymentConfig operation.

See: createDeploymentConfigResponse smart constructor.

Response Lenses

cdcrsDeploymentConfigId :: Lens' CreateDeploymentConfigResponse (Maybe Text)

A unique deployment configuration ID.

cdcrsResponseStatus :: Lens' CreateDeploymentConfigResponse Int

- | The response status code.