Network.AWS.CodeDeploy.CreateDeploymentGroup

Description

Creates a deployment group to which application revisions will be deployed.

Creating a Request

createDeploymentGroup

Arguments

:: Text
cdgApplicationName

-> Text
cdgDeploymentGroupName

-> Text
cdgServiceRoleARN

-> CreateDeploymentGroup

Creates a value of CreateDeploymentGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdgEc2TagSet - Information about groups of tags applied to EC2 instances. The deployment group will include only EC2 instances identified by all the tag groups. Cannot be used in the same call as ec2TagFilters.

cdgDeploymentConfigName - If specified, the deployment configuration name can be either one of the predefined configurations provided with AWS CodeDeploy or a custom deployment configuration that you create by calling the create deployment configuration operation. CodeDeployDefault.OneAtATime is the default deployment configuration. It is used if a configuration isn't specified for the deployment or the deployment group. For more information about the predefined deployment configurations in AWS CodeDeploy, see Working with Deployment Groups in AWS CodeDeploy in the AWS CodeDeploy User Guide.

cdgOnPremisesTagSet - Information about groups of tags applied to on-premises instances. The deployment group will include only on-premises instances identified by all the tag groups. Cannot be used in the same call as onPremisesInstanceTagFilters.

cdgEc2TagFilters - The Amazon EC2 tags on which to filter. The deployment group will include EC2 instances with any of the specified tags. Cannot be used in the same call as ec2TagSet.

cdgBlueGreenDeploymentConfiguration - Information about blue/green deployment options for a deployment group.

cdgLoadBalancerInfo - Information about the load balancer used in a deployment.

cdgOnPremisesInstanceTagFilters - The on-premises instance tags on which to filter. The deployment group will include on-premises instances with any of the specified tags. Cannot be used in the same call as OnPremisesTagSet.

cdgAlarmConfiguration - Information to add about Amazon CloudWatch alarms when the deployment group is created.

cdgTriggerConfigurations - Information about triggers to create when the deployment group is created. For examples, see Create a Trigger for an AWS CodeDeploy Event in the AWS CodeDeploy User Guide.

cdgAutoScalingGroups - A list of associated Auto Scaling groups.

cdgDeploymentStyle - Information about the type of deployment, in-place or blue/green, that you want to run and whether to route deployment traffic behind a load balancer.

cdgAutoRollbackConfiguration - Configuration information for an automatic rollback that is added when a deployment group is created.

cdgApplicationName - The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.

cdgDeploymentGroupName - The name of a new deployment group for the specified application.

cdgServiceRoleARN - A service role ARN that allows AWS CodeDeploy to act on the user's behalf when interacting with AWS services.

data CreateDeploymentGroup

Represents the input of a CreateDeploymentGroup operation.

See: createDeploymentGroup smart constructor.

Request Lenses

cdgEc2TagSet :: Lens' CreateDeploymentGroup (Maybe EC2TagSet)

Information about groups of tags applied to EC2 instances. The deployment group will include only EC2 instances identified by all the tag groups. Cannot be used in the same call as ec2TagFilters.

cdgDeploymentConfigName :: Lens' CreateDeploymentGroup (Maybe Text)

If specified, the deployment configuration name can be either one of the predefined configurations provided with AWS CodeDeploy or a custom deployment configuration that you create by calling the create deployment configuration operation. CodeDeployDefault.OneAtATime is the default deployment configuration. It is used if a configuration isn't specified for the deployment or the deployment group. For more information about the predefined deployment configurations in AWS CodeDeploy, see Working with Deployment Groups in AWS CodeDeploy in the AWS CodeDeploy User Guide.

cdgOnPremisesTagSet :: Lens' CreateDeploymentGroup (Maybe OnPremisesTagSet)

Information about groups of tags applied to on-premises instances. The deployment group will include only on-premises instances identified by all the tag groups. Cannot be used in the same call as onPremisesInstanceTagFilters.

cdgEc2TagFilters :: Lens' CreateDeploymentGroup [EC2TagFilter]

The Amazon EC2 tags on which to filter. The deployment group will include EC2 instances with any of the specified tags. Cannot be used in the same call as ec2TagSet.

cdgBlueGreenDeploymentConfiguration :: Lens' CreateDeploymentGroup (Maybe BlueGreenDeploymentConfiguration)

Information about blue/green deployment options for a deployment group.

cdgLoadBalancerInfo :: Lens' CreateDeploymentGroup (Maybe LoadBalancerInfo)

Information about the load balancer used in a deployment.

cdgOnPremisesInstanceTagFilters :: Lens' CreateDeploymentGroup [TagFilter]

The on-premises instance tags on which to filter. The deployment group will include on-premises instances with any of the specified tags. Cannot be used in the same call as OnPremisesTagSet.

cdgAlarmConfiguration :: Lens' CreateDeploymentGroup (Maybe AlarmConfiguration)

Information to add about Amazon CloudWatch alarms when the deployment group is created.

cdgTriggerConfigurations :: Lens' CreateDeploymentGroup [TriggerConfig]

Information about triggers to create when the deployment group is created. For examples, see Create a Trigger for an AWS CodeDeploy Event in the AWS CodeDeploy User Guide.

cdgAutoScalingGroups :: Lens' CreateDeploymentGroup [Text]

A list of associated Auto Scaling groups.

cdgDeploymentStyle :: Lens' CreateDeploymentGroup (Maybe DeploymentStyle)

Information about the type of deployment, in-place or blue/green, that you want to run and whether to route deployment traffic behind a load balancer.

cdgAutoRollbackConfiguration :: Lens' CreateDeploymentGroup (Maybe AutoRollbackConfiguration)

Configuration information for an automatic rollback that is added when a deployment group is created.

cdgApplicationName :: Lens' CreateDeploymentGroup Text

The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.

cdgDeploymentGroupName :: Lens' CreateDeploymentGroup Text

The name of a new deployment group for the specified application.

cdgServiceRoleARN :: Lens' CreateDeploymentGroup Text

A service role ARN that allows AWS CodeDeploy to act on the user's behalf when interacting with AWS services.

Destructuring the Response

createDeploymentGroupResponse

Arguments

:: Int
cdgrsResponseStatus

-> CreateDeploymentGroupResponse

Creates a value of CreateDeploymentGroupResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdgrsDeploymentGroupId - A unique deployment group ID.

cdgrsResponseStatus - -- | The response status code.

data CreateDeploymentGroupResponse

Represents the output of a CreateDeploymentGroup operation.

See: createDeploymentGroupResponse smart constructor.

Response Lenses

cdgrsDeploymentGroupId :: Lens' CreateDeploymentGroupResponse (Maybe Text)

A unique deployment group ID.

cdgrsResponseStatus :: Lens' CreateDeploymentGroupResponse Int

- | The response status code.