Network.AWS.CodeDeploy.DeleteDeploymentGroup

Description

Deletes a deployment group.

Creating a Request

deleteDeploymentGroup

Arguments

:: Text
ddgApplicationName

-> Text
ddgDeploymentGroupName

-> DeleteDeploymentGroup

Creates a value of DeleteDeploymentGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddgApplicationName - The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.

ddgDeploymentGroupName - The name of an existing deployment group for the specified application.

data DeleteDeploymentGroup

Represents the input of a DeleteDeploymentGroup operation.

See: deleteDeploymentGroup smart constructor.

Request Lenses

ddgApplicationName :: Lens' DeleteDeploymentGroup Text

The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.

ddgDeploymentGroupName :: Lens' DeleteDeploymentGroup Text

The name of an existing deployment group for the specified application.

Destructuring the Response

deleteDeploymentGroupResponse

Arguments

:: Int
ddgrsResponseStatus

-> DeleteDeploymentGroupResponse

Creates a value of DeleteDeploymentGroupResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddgrsHooksNotCleanedUp - If the output contains no data, and the corresponding deployment group contained at least one Auto Scaling group, AWS CodeDeploy successfully removed all corresponding Auto Scaling lifecycle event hooks from the Amazon EC2 instances in the Auto Scaling group. If the output contains data, AWS CodeDeploy could not remove some Auto Scaling lifecycle event hooks from the Amazon EC2 instances in the Auto Scaling group.

ddgrsResponseStatus - -- | The response status code.

data DeleteDeploymentGroupResponse

Represents the output of a DeleteDeploymentGroup operation.

See: deleteDeploymentGroupResponse smart constructor.

Response Lenses

ddgrsHooksNotCleanedUp :: Lens' DeleteDeploymentGroupResponse [AutoScalingGroup]

If the output contains no data, and the corresponding deployment group contained at least one Auto Scaling group, AWS CodeDeploy successfully removed all corresponding Auto Scaling lifecycle event hooks from the Amazon EC2 instances in the Auto Scaling group. If the output contains data, AWS CodeDeploy could not remove some Auto Scaling lifecycle event hooks from the Amazon EC2 instances in the Auto Scaling group.

ddgrsResponseStatus :: Lens' DeleteDeploymentGroupResponse Int

- | The response status code.