Network.AWS.CodeDeploy.GetApplicationRevision

Description

Gets information about an application revision.

Creating a Request

getApplicationRevision

Arguments

:: Text
garApplicationName

-> RevisionLocation
garRevision

-> GetApplicationRevision

Creates a value of GetApplicationRevision with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

garApplicationName - The name of the application that corresponds to the revision.

garRevision - Information about the application revision to get, including type and location.

data GetApplicationRevision

Represents the input of a GetApplicationRevision operation.

See: getApplicationRevision smart constructor.

Request Lenses

garApplicationName :: Lens' GetApplicationRevision Text

The name of the application that corresponds to the revision.

garRevision :: Lens' GetApplicationRevision RevisionLocation

Information about the application revision to get, including type and location.

Destructuring the Response

getApplicationRevisionResponse

Arguments

:: Int
garrsResponseStatus

-> GetApplicationRevisionResponse

Creates a value of GetApplicationRevisionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

garrsApplicationName - The name of the application that corresponds to the revision.

garrsRevisionInfo - General information about the revision.

garrsRevision - Additional information about the revision, including type and location.

garrsResponseStatus - -- | The response status code.

data GetApplicationRevisionResponse

Represents the output of a GetApplicationRevision operation.

See: getApplicationRevisionResponse smart constructor.

Response Lenses

garrsApplicationName :: Lens' GetApplicationRevisionResponse (Maybe Text)

The name of the application that corresponds to the revision.

garrsRevisionInfo :: Lens' GetApplicationRevisionResponse (Maybe GenericRevisionInfo)

General information about the revision.

garrsRevision :: Lens' GetApplicationRevisionResponse (Maybe RevisionLocation)

Additional information about the revision, including type and location.

garrsResponseStatus :: Lens' GetApplicationRevisionResponse Int

- | The response status code.