Network.AWS.CodeDeploy.ListApplicationRevisions

Description

Lists information about revisions for an application.

This operation returns paginated results.

Creating a Request

listApplicationRevisions

Arguments

:: Text
larApplicationName

-> ListApplicationRevisions

Creates a value of ListApplicationRevisions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

larS3KeyPrefix - A key prefix for the set of Amazon S3 objects to limit the search for revisions.

larDeployed - Whether to list revisions based on whether the revision is the target revision of an deployment group: * include: List revisions that are target revisions of a deployment group. * exclude: Do not list revisions that are target revisions of a deployment group. * ignore: List all revisions.

larSortOrder - The order in which to sort the list results: * ascending: ascending order. * descending: descending order. If not specified, the results will be sorted in ascending order. If set to null, the results will be sorted in an arbitrary order.

larNextToken - An identifier returned from the previous list application revisions call. It can be used to return the next set of applications in the list.

larS3Bucket - An Amazon S3 bucket name to limit the search for revisions. If set to null, all of the user's buckets will be searched.

larSortBy - The column name to use to sort the list results: * registerTime: Sort by the time the revisions were registered with AWS CodeDeploy. * firstUsedTime: Sort by the time the revisions were first used in a deployment. * lastUsedTime: Sort by the time the revisions were last used in a deployment. If not specified or set to null, the results will be returned in an arbitrary order.

larApplicationName - The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.

data ListApplicationRevisions

Represents the input of a ListApplicationRevisions operation.

See: listApplicationRevisions smart constructor.

Request Lenses

larS3KeyPrefix :: Lens' ListApplicationRevisions (Maybe Text)

A key prefix for the set of Amazon S3 objects to limit the search for revisions.

larDeployed :: Lens' ListApplicationRevisions (Maybe ListStateFilterAction)

Whether to list revisions based on whether the revision is the target revision of an deployment group: * include: List revisions that are target revisions of a deployment group. * exclude: Do not list revisions that are target revisions of a deployment group. * ignore: List all revisions.

larSortOrder :: Lens' ListApplicationRevisions (Maybe SortOrder)

The order in which to sort the list results: * ascending: ascending order. * descending: descending order. If not specified, the results will be sorted in ascending order. If set to null, the results will be sorted in an arbitrary order.

larNextToken :: Lens' ListApplicationRevisions (Maybe Text)

An identifier returned from the previous list application revisions call. It can be used to return the next set of applications in the list.

larS3Bucket :: Lens' ListApplicationRevisions (Maybe Text)

An Amazon S3 bucket name to limit the search for revisions. If set to null, all of the user's buckets will be searched.

larSortBy :: Lens' ListApplicationRevisions (Maybe ApplicationRevisionSortBy)

The column name to use to sort the list results: * registerTime: Sort by the time the revisions were registered with AWS CodeDeploy. * firstUsedTime: Sort by the time the revisions were first used in a deployment. * lastUsedTime: Sort by the time the revisions were last used in a deployment. If not specified or set to null, the results will be returned in an arbitrary order.

larApplicationName :: Lens' ListApplicationRevisions Text

The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.

Destructuring the Response

listApplicationRevisionsResponse

Arguments

:: Int
larrsResponseStatus

-> ListApplicationRevisionsResponse

Creates a value of ListApplicationRevisionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

larrsNextToken - If a large amount of information is returned, an identifier will also be returned. It can be used in a subsequent list application revisions call to return the next set of application revisions in the list.

larrsRevisions - A list of locations that contain the matching revisions.

larrsResponseStatus - -- | The response status code.

data ListApplicationRevisionsResponse

Represents the output of a ListApplicationRevisions operation.

See: listApplicationRevisionsResponse smart constructor.

Response Lenses

larrsNextToken :: Lens' ListApplicationRevisionsResponse (Maybe Text)

If a large amount of information is returned, an identifier will also be returned. It can be used in a subsequent list application revisions call to return the next set of application revisions in the list.

larrsRevisions :: Lens' ListApplicationRevisionsResponse [RevisionLocation]

A list of locations that contain the matching revisions.

larrsResponseStatus :: Lens' ListApplicationRevisionsResponse Int

- | The response status code.