Network.AWS.CodeDeploy.ListDeploymentGroups

Description

Lists the deployment groups for an application registered with the applicable IAM user or AWS account.

This operation returns paginated results.

Creating a Request

listDeploymentGroups

Arguments

:: Text
ldgApplicationName

-> ListDeploymentGroups

Creates a value of ListDeploymentGroups with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ldgNextToken - An identifier returned from the previous list deployment groups call. It can be used to return the next set of deployment groups in the list.

ldgApplicationName - The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.

data ListDeploymentGroups

Represents the input of a ListDeploymentGroups operation.

See: listDeploymentGroups smart constructor.

Request Lenses

ldgNextToken :: Lens' ListDeploymentGroups (Maybe Text)

An identifier returned from the previous list deployment groups call. It can be used to return the next set of deployment groups in the list.

ldgApplicationName :: Lens' ListDeploymentGroups Text

The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.

Destructuring the Response

listDeploymentGroupsResponse

Arguments

:: Int
ldgrsResponseStatus

-> ListDeploymentGroupsResponse

Creates a value of ListDeploymentGroupsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ldgrsNextToken - If a large amount of information is returned, an identifier is also returned. It can be used in a subsequent list deployment groups call to return the next set of deployment groups in the list.

ldgrsApplicationName - The application name.

ldgrsDeploymentGroups - A list of corresponding deployment group names.

ldgrsResponseStatus - -- | The response status code.

data ListDeploymentGroupsResponse

Represents the output of a ListDeploymentGroups operation.

See: listDeploymentGroupsResponse smart constructor.

Response Lenses

ldgrsNextToken :: Lens' ListDeploymentGroupsResponse (Maybe Text)

If a large amount of information is returned, an identifier is also returned. It can be used in a subsequent list deployment groups call to return the next set of deployment groups in the list.

ldgrsApplicationName :: Lens' ListDeploymentGroupsResponse (Maybe Text)

The application name.

ldgrsDeploymentGroups :: Lens' ListDeploymentGroupsResponse [Text]

A list of corresponding deployment group names.

ldgrsResponseStatus :: Lens' ListDeploymentGroupsResponse Int

- | The response status code.