Network.AWS.CodeDeploy.ListDeployments

Description

Lists the deployments in a deployment group for an application registered with the applicable IAM user or AWS account.

This operation returns paginated results.

Creating a Request

listDeployments :: ListDeployments

Creates a value of ListDeployments with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ldCreateTimeRange - A time range (start and end) for returning a subset of the list of deployments.

ldNextToken - An identifier returned from the previous list deployments call. It can be used to return the next set of deployments in the list.

ldIncludeOnlyStatuses - A subset of deployments to list by status: * Created: Include created deployments in the resulting list. * Queued: Include queued deployments in the resulting list. * In Progress: Include in-progress deployments in the resulting list. * Succeeded: Include successful deployments in the resulting list. * Failed: Include failed deployments in the resulting list. * Stopped: Include stopped deployments in the resulting list.

ldApplicationName - The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.

ldDeploymentGroupName - The name of an existing deployment group for the specified application.

data ListDeployments

Represents the input of a ListDeployments operation.

See: listDeployments smart constructor.

Request Lenses

ldCreateTimeRange :: Lens' ListDeployments (Maybe TimeRange)

A time range (start and end) for returning a subset of the list of deployments.

ldNextToken :: Lens' ListDeployments (Maybe Text)

An identifier returned from the previous list deployments call. It can be used to return the next set of deployments in the list.

ldIncludeOnlyStatuses :: Lens' ListDeployments [DeploymentStatus]

A subset of deployments to list by status: * Created: Include created deployments in the resulting list. * Queued: Include queued deployments in the resulting list. * In Progress: Include in-progress deployments in the resulting list. * Succeeded: Include successful deployments in the resulting list. * Failed: Include failed deployments in the resulting list. * Stopped: Include stopped deployments in the resulting list.

ldApplicationName :: Lens' ListDeployments (Maybe Text)

The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.

ldDeploymentGroupName :: Lens' ListDeployments (Maybe Text)

The name of an existing deployment group for the specified application.

Destructuring the Response

listDeploymentsResponse

Arguments

:: Int
ldrsResponseStatus

-> ListDeploymentsResponse

Creates a value of ListDeploymentsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ldrsNextToken - If a large amount of information is returned, an identifier is also returned. It can be used in a subsequent list deployments call to return the next set of deployments in the list.

ldrsDeployments - A list of deployment IDs.

ldrsResponseStatus - -- | The response status code.

data ListDeploymentsResponse

Represents the output of a ListDeployments operation.

See: listDeploymentsResponse smart constructor.

Response Lenses

ldrsNextToken :: Lens' ListDeploymentsResponse (Maybe Text)

If a large amount of information is returned, an identifier is also returned. It can be used in a subsequent list deployments call to return the next set of deployments in the list.

ldrsDeployments :: Lens' ListDeploymentsResponse [Text]

A list of deployment IDs.

ldrsResponseStatus :: Lens' ListDeploymentsResponse Int

- | The response status code.