Network.AWS.CodeDeploy.UpdateDeploymentGroup

Description

Changes information about a deployment group.

Creating a Request

updateDeploymentGroup

Arguments

:: Text
udgApplicationName

-> Text
udgCurrentDeploymentGroupName

-> UpdateDeploymentGroup

Creates a value of UpdateDeploymentGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

udgServiceRoleARN - A replacement ARN for the service role, if you want to change it.

udgEc2TagSet - Information about groups of tags applied to on-premises instances. The deployment group will include only EC2 instances identified by all the tag groups.

udgDeploymentConfigName - The replacement deployment configuration name to use, if you want to change it.

udgOnPremisesTagSet - Information about an on-premises instance tag set. The deployment group will include only on-premises instances identified by all the tag groups.

udgNewDeploymentGroupName - The new name of the deployment group, if you want to change it.

udgEc2TagFilters - The replacement set of Amazon EC2 tags on which to filter, if you want to change them. To keep the existing tags, enter their names. To remove tags, do not enter any tag names.

udgBlueGreenDeploymentConfiguration - Information about blue/green deployment options for a deployment group.

udgLoadBalancerInfo - Information about the load balancer used in a deployment.

udgOnPremisesInstanceTagFilters - The replacement set of on-premises instance tags on which to filter, if you want to change them. To keep the existing tags, enter their names. To remove tags, do not enter any tag names.

udgAlarmConfiguration - Information to add or change about Amazon CloudWatch alarms when the deployment group is updated.

udgTriggerConfigurations - Information about triggers to change when the deployment group is updated. For examples, see Modify Triggers in an AWS CodeDeploy Deployment Group in the AWS CodeDeploy User Guide.

udgAutoScalingGroups - The replacement list of Auto Scaling groups to be included in the deployment group, if you want to change them. To keep the Auto Scaling groups, enter their names. To remove Auto Scaling groups, do not enter any Auto Scaling group names.

udgDeploymentStyle - Information about the type of deployment, either in-place or blue/green, you want to run and whether to route deployment traffic behind a load balancer.

udgAutoRollbackConfiguration - Information for an automatic rollback configuration that is added or changed when a deployment group is updated.

udgApplicationName - The application name corresponding to the deployment group to update.

udgCurrentDeploymentGroupName - The current name of the deployment group.

data UpdateDeploymentGroup

Represents the input of an UpdateDeploymentGroup operation.

See: updateDeploymentGroup smart constructor.

Request Lenses

udgServiceRoleARN :: Lens' UpdateDeploymentGroup (Maybe Text)

A replacement ARN for the service role, if you want to change it.

udgEc2TagSet :: Lens' UpdateDeploymentGroup (Maybe EC2TagSet)

Information about groups of tags applied to on-premises instances. The deployment group will include only EC2 instances identified by all the tag groups.

udgDeploymentConfigName :: Lens' UpdateDeploymentGroup (Maybe Text)

The replacement deployment configuration name to use, if you want to change it.

udgOnPremisesTagSet :: Lens' UpdateDeploymentGroup (Maybe OnPremisesTagSet)

Information about an on-premises instance tag set. The deployment group will include only on-premises instances identified by all the tag groups.

udgNewDeploymentGroupName :: Lens' UpdateDeploymentGroup (Maybe Text)

The new name of the deployment group, if you want to change it.

udgEc2TagFilters :: Lens' UpdateDeploymentGroup [EC2TagFilter]

The replacement set of Amazon EC2 tags on which to filter, if you want to change them. To keep the existing tags, enter their names. To remove tags, do not enter any tag names.

udgBlueGreenDeploymentConfiguration :: Lens' UpdateDeploymentGroup (Maybe BlueGreenDeploymentConfiguration)

Information about blue/green deployment options for a deployment group.

udgLoadBalancerInfo :: Lens' UpdateDeploymentGroup (Maybe LoadBalancerInfo)

Information about the load balancer used in a deployment.

udgOnPremisesInstanceTagFilters :: Lens' UpdateDeploymentGroup [TagFilter]

The replacement set of on-premises instance tags on which to filter, if you want to change them. To keep the existing tags, enter their names. To remove tags, do not enter any tag names.

udgAlarmConfiguration :: Lens' UpdateDeploymentGroup (Maybe AlarmConfiguration)

Information to add or change about Amazon CloudWatch alarms when the deployment group is updated.

udgTriggerConfigurations :: Lens' UpdateDeploymentGroup [TriggerConfig]

Information about triggers to change when the deployment group is updated. For examples, see Modify Triggers in an AWS CodeDeploy Deployment Group in the AWS CodeDeploy User Guide.

udgAutoScalingGroups :: Lens' UpdateDeploymentGroup [Text]

The replacement list of Auto Scaling groups to be included in the deployment group, if you want to change them. To keep the Auto Scaling groups, enter their names. To remove Auto Scaling groups, do not enter any Auto Scaling group names.

udgDeploymentStyle :: Lens' UpdateDeploymentGroup (Maybe DeploymentStyle)

Information about the type of deployment, either in-place or blue/green, you want to run and whether to route deployment traffic behind a load balancer.

udgAutoRollbackConfiguration :: Lens' UpdateDeploymentGroup (Maybe AutoRollbackConfiguration)

Information for an automatic rollback configuration that is added or changed when a deployment group is updated.

udgApplicationName :: Lens' UpdateDeploymentGroup Text

The application name corresponding to the deployment group to update.

udgCurrentDeploymentGroupName :: Lens' UpdateDeploymentGroup Text

The current name of the deployment group.

Destructuring the Response

updateDeploymentGroupResponse

Arguments

:: Int
udgrsResponseStatus

-> UpdateDeploymentGroupResponse

Creates a value of UpdateDeploymentGroupResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

udgrsHooksNotCleanedUp - If the output contains no data, and the corresponding deployment group contained at least one Auto Scaling group, AWS CodeDeploy successfully removed all corresponding Auto Scaling lifecycle event hooks from the AWS account. If the output contains data, AWS CodeDeploy could not remove some Auto Scaling lifecycle event hooks from the AWS account.

udgrsResponseStatus - -- | The response status code.

data UpdateDeploymentGroupResponse

Represents the output of an UpdateDeploymentGroup operation.

See: updateDeploymentGroupResponse smart constructor.

Response Lenses

udgrsHooksNotCleanedUp :: Lens' UpdateDeploymentGroupResponse [AutoScalingGroup]

If the output contains no data, and the corresponding deployment group contained at least one Auto Scaling group, AWS CodeDeploy successfully removed all corresponding Auto Scaling lifecycle event hooks from the AWS account. If the output contains data, AWS CodeDeploy could not remove some Auto Scaling lifecycle event hooks from the AWS account.

udgrsResponseStatus :: Lens' UpdateDeploymentGroupResponse Int

- | The response status code.